@@ -19,8 +19,8 @@
   "homepage": "https://github.com/peterpwang/workers-schedule#readme",
   "dependencies": {
     "net": "^1.0.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0",
     "react-router-dom": "^4.3.1",
     "rest": "^1.3.1",
     "sockjs-client": "^1.0.3",
