@@ -5,15 +5,45 @@ import { BrowserRouter, Switch, Route, Link } from "react-router-dom";
 import LoginApp from './login'; 
 import ScheduleApp from './schedule'; 
 import UserApp from './user'; 
+import LoginContext from './logincontext';
 
 class App extends React.Component {
+	
+	constructor(props) {
+		super(props);
+
+		this.toggleLogin = this.toggleLogin.bind(this);		
+		this.state = {
+            authorization: undefined,
+			toggleLogin: this.toggleLogin
+		};
+	}
+	
 	render() {
 		return (
 		  <div>
-			<Header />
-			<Main />
+			<Switch>
+			  <Route exact path='/' render={() => {
+				  return (<LoginContext.Provider value={this.state}><LoginApp/></LoginContext.Provider>);
+				}
+			  }/>
+			  <Route path='/users' render={() => {
+				  return (<div><Header/><main><UserApp/></main></div>);
+				}
+			  }/>
+			  <Route path='/schedules' render={() => {
+				  return (<div><Header/><main><ScheduleApp/></main></div>);
+				}
+			  }/>
+			</Switch>
 		  </div>
 		)
+	}
+	
+	toggleLogin(newAuthorization) {
+		this.setState(state => ({
+			authorization: newAuthorization
+        }));
 	}
 }
 
@@ -35,20 +65,6 @@ class Header extends React.Component {
 				  </ul>
 				</nav>
 			</header>
-		)
-	}
-}
-
-class Main extends React.Component {
-	render() {
-		return (
-			<main>
-				<Switch>
-				  <Route exact path='/' render={() => <LoginApp/>}/>
-				  <Route path='/users' render={() => <UserApp/>}/>
-				  <Route path='/schedules' render={() => <ScheduleApp/>}/>
-				</Switch>
-			</main>
 		)
 	}
 }
